refactor: move music cards into subdir
