+++ components/music/cards/index.tsx
@@ -1,0 +1,4 @@
+export { default as Album } from "./Album";
+export { default as Artist } from "./Artist";
+export { default as Playlist } from "./Playlist";
+export { default as Track } from "./Track";
